requirements.txt
pydantic
pytest
pyarrow
pandas
amazon.ion